+++ src/main/java/com/example/redunm/CartDataInitializer.java
@@ -3,10 +3,10 @@ package com.example.redunm;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
-import com.example.redunm.entity.Cart;
-import com.example.redunm.entity.CartItem;
+import com.example.redunm.cart.Cart;
+import com.example.redunm.cart.CartItem;
 import com.example.redunm.modellist.DataModel;
-import com.example.redunm.modellist.CartRepository;
+import com.example.redunm.cart.CartRepository;
 import com.example.redunm.modellist.DataModelRepository;
 
 import java.util.Collections;
@@ -28,7 +28,7 @@ public class CartDataInitializer implements CommandLineRunner {
     private void createMockData() {
         // 1. 가상 DataModel 생성 및 저장
         DataModel model = new DataModel();
-        model.setName("della");
+        model.setName("yoyak");
         model.setTag("meeting");
         model.setPrice(150);
         dataModelRepository.save(model);
