--- src/main/java/com/example/redunm/dto/ModelImgDto.java
@@ -1,4 +1,0 @@
-package com.example.redunm.dto;
-
-public class ModelImgDto {
-}
